bugfix: off-by-one in ConstantPool GetSize() and serializer loop
After a previous change that made it so the ConstantPool stores nullptr
at index 0 to ensure correct 1 based indexing, the GetSize() function weren't
updated to account for this which resulted in it being one greater than it
should.

Additionally the serializers SerializeConstantPool() function weren't updated
to iterate from 1 as it should.

This commit fixes those issues.